add some conditions in Javadoc publisher

+++ .script/publish_javadoc.sh
@@ -2,11 +2,16 @@
 
 repo_name="atlanmod/NeoEMF"
 jdk="oraclejdk8"
-
-path="apidocs"
-temp_path="$HOME/$path"
+path=apidocs
+temp_path=$HOME/${path}
 
 if [ "$TRAVIS_REPO_SLUG" = "$repo_name" ] && [ "$TRAVIS_JDK_VERSION" = "$jdk" ] && [ "$TRAVIS_PULL_REQUEST" = "false" ] && [ "$TRAVIS_BRANCH" = "master" ]; then
+
+    if ! [ -d target/site/apidocs ]; then
+        echo -e "No new Javadoc.\n"
+        exit
+    fi
+
     echo -e "Publishing Javadoc...\n"
 
     # Copy the built javadoc
@@ -20,7 +25,11 @@ if [ "$TRAVIS_REPO_SLUG" = "$repo_name" ] && [ "$TRAVIS_JDK_VERSION" = "$jdk" ] 
 
     # Update the javadoc in 'gh-pages' branch
     cd gh-pages
-    git rm --quiet -rf ${path}
+
+    if [ -d "$path" ]; then
+        git rm --quiet -rf ${path}
+    fi
+
     cp -Rf ${temp_path} ${path}
 
     # Commit modifications
